+++ src/services/github.service.ts
@@ -1,5 +1,20 @@
 const GITHUB_API_URL = 'https://api.github.com';
 
+export type GitHubRepo = {
+  id: number;
+  full_name: string;
+  created_at: string;
+  stargazers_count: number;
+  language: string;
+  forks_count: number;
+  open_issues_count: number;
+  license: {
+    name: string;
+  } | null;
+  lastCommitAt: string | null;
+  isFavorited: boolean;
+};
+
 export const getRepository = async ({
   username,
   repositoryName,
