+++ src/utils/constants.ts
@@ -1,0 +1,1 @@
+export const MINIMUM_REPO_NAME_LENGTH = 1;
